test: - move describe with ListItem tests to ListItem.test.ts
